refactor: move production Clock/Sleeper impls out of stella-core (fixes #67)
stella-core carried the production implementations of its own time
ports — SystemClock (ports.rs) and TokioSleeper (retry.rs) — and
Engine::new wired TokioSleeper internally, breaking the crate's
ports-not-concretions boundary and pinning a production tokio::time
call inside the engine.

- stella-core now exports only the Clock and Sleeper traits;
  Engine::with_sleeper is the sole constructor (Engine::new, which
  baked in TokioSleeper, is removed).
- The production impls move to stella-cli/src/runtime.rs — the CLI
  binary is the only workspace crate that constructed them (fleet,
  pipeline, and context already take their time ports by injection
  or define their own), so no new adapter crate is warranted.
- SystemClock's unit tests move with it; retry/driver/goal tests
  already used local no-op sleepers and are unchanged.
- tokio stays in stella-core for tokio::sync::mpsc (driver/goal event
  channels); its production tokio::time use is gone.
- Silence one pre-existing dead-code warning and one clippy map_entry
  warning (both from #94, unrelated) to keep the workspace warning-free.

Fixes #67
