+++ stella-cli/src/main.rs
@@ -27,6 +27,7 @@ mod fleet_cmd;
 mod interactive;
 mod memory;
 mod ocp;
+mod runtime;
 mod settings;
 mod stats;
 mod tui;
